gather all test to single file
