+++ hardfloat/tests/src/ValExec_CompareRecFN.scala
@@ -132,47 +132,3 @@ class ValExec_CompareRecFN_eq(expWidth: Int, sigWidth: Int) extends Module
         (io.actual.out === io.expected.out) &&
         (io.actual.exceptionFlags === io.expected.exceptionFlags)
 }
-
-class CompareRecFNSpec extends FMATester {
-    def test(f: Int, fn: String): Seq[String] = {
-        val generator = fn match {
-            case "lt" => () => new ValExec_CompareRecFN_lt(exp(f), sig(f))
-            case "le" => () => new ValExec_CompareRecFN_le(exp(f), sig(f))
-            case "eq" => () => new ValExec_CompareRecFN_eq(exp(f), sig(f))
-        }
-        test(
-            s"CompareRecF${f}_$fn",
-            generator,
-            "CompareRecFN.cpp",
-            Seq(Seq(s"f${f}_${fn}"))
-        )
-    }
-
-    "CompareRecF16_lt" should "pass" in {
-        check(test(16, "lt"))
-    }
-    "CompareRecF32_lt" should "pass" in {
-        check(test(32, "lt"))
-    }
-    "CompareRecF64_lt" should "pass" in {
-        check(test(64, "lt"))
-    }
-    "CompareRecF16_le" should "pass" in {
-        check(test(16, "le"))
-    }
-    "CompareRecF32_le" should "pass" in {
-        check(test(32, "le"))
-    }
-    "CompareRecF64_le" should "pass" in {
-        check(test(64, "le"))
-    }
-    "CompareRecF16_eq" should "pass" in {
-        check(test(16, "eq"))
-    }
-    "CompareRecF32_eq" should "pass" in {
-        check(test(32, "eq"))
-    }
-    "CompareRecF64_eq" should "pass" in {
-        check(test(64, "eq"))
-  }
-}
