+++ hardfloat/tests/src/ValExec_DivSqrtRecFN_small.scala
@@ -183,40 +183,3 @@ class
         equivRecFN(expWidth, sigWidth, io.actual.out, io.expected.recOut) &&
         (io.actual.exceptionFlags === io.expected.exceptionFlags)
 }
-
-class DivSqrtRecFn_smallSpec extends FMATester {
-    def test(f: Int, fn: String): Seq[String] = {
-        def generator(options: Int) = fn match {
-            case "div" => () => new ValExec_DivSqrtRecFN_small_div(exp(f), sig(f), options)
-            case "sqrt" => () => new ValExec_DivSqrtRecFN_small_sqrt(exp(f), sig(f), options)
-        }
-        test(
-            s"DivSqrtRecF${f}_small_${fn}",
-            generator(0),
-            (if (fn == "sqrt") Seq("-level2") else Seq.empty) ++ Seq(s"f${f}_${fn}")
-        )
-        test(
-            s"DivSqrtRecF${f}_small_${fn}",
-            generator(consts.divSqrtOpt_twoBitsPerCycle),
-            (if (fn == "sqrt") Seq("-level2") else Seq.empty) ++ Seq(s"f${f}_${fn}")
-        )
-    }
-    "DivSqrtRecF16_small_div" should "pass" in {
-        check(test(16, "div"))
-    }
-    "DivSqrtRecF32_small_div" should "pass" in {
-        check(test(32, "div"))
-    }
-    "DivSqrtRecF64_small_div" should "pass" in {
-        check(test(64, "div"))
-    }
-    "DivSqrtRecF16_small_sqrt" should "pass" in {
-        check(test(16, "sqrt"))
-    }
-    "DivSqrtRecF32_small_sqrt" should "pass" in {
-        check(test(32, "sqrt"))
-    }
-    "DivSqrtRecF64_small_sqrt" should "pass" in {
-        check(test(64, "sqrt"))
-    }
-}
