+++ hardfloat/tests/src/ValExec_MulAddRecFN.scala
@@ -166,48 +166,3 @@ class ValExec_MulAddRecFN_mul(expWidth: Int, sigWidth: Int) extends Module
         equivRecFN(expWidth, sigWidth, io.actual.out, io.expected.recOut) &&
         (io.actual.exceptionFlags === io.expected.exceptionFlags)
 }
-
-class MulAddRecFNSpec extends FMATester {
-    def test(f: Int, fn: String): Seq[String] = {
-        test(
-            s"MulAddRecF${f}${fn match {
-                case "add" => "_add"
-                case "mul" => "_mul"
-                case "mulAdd" => ""
-            }}",
-            () => fn match {
-                case "add" => new ValExec_MulAddRecFN_add(exp(f), sig(f))
-                case "mul" => new ValExec_MulAddRecFN_mul(exp(f), sig(f))
-                case "mulAdd" => new ValExec_MulAddRecFN(exp(f), sig(f))
-            },
-            Seq(s"f${f}_${fn}")
-        )
-    }
-    "MulAddRecF16" should "pass" in {
-        check(test(16, "mulAdd"))
-    }
-    "MulAddRecF32" should "pass" in {
-        check(test(32, "mulAdd"))
-    }
-    "MulAddRecF64" should "pass" in {
-        check(test(64, "mulAdd"))
-    }
-    "MulAddRecF16_add" should "pass" in {
-        check(test(16, "add"))
-    }
-    "MulAddRecF32_add" should "pass" in {
-        check(test(32, "add"))
-    }
-    "MulAddRecF64_add" should "pass" in {
-        check(test(64, "add"))
-    }
-    "MulAddRecF16_mul" should "pass" in {
-        check(test(16, "mul"))
-    }
-    "MulAddRecF32_mul" should "pass" in {
-        check(test(32, "mul"))
-    }
-    "MulAddRecF64_mul" should "pass" in {
-        check(test(64, "mul"))
-    }
-}
