+++ hardfloat/tests/src/ValExec_RecFNToIN.scala
@@ -119,70 +119,3 @@ class
         (io.actual.out === io.expected.out) &&
         (io.actual.exceptionFlags === io.expected.exceptionFlags)
 }
-class RecFNToUINSpec extends FMATester {
-    def test(f: Int, i: Int): Seq[String] = {
-        val (softfloatArgs, dutArgs) = roundings.map { case (s, d) =>
-            (s +: Seq("-exact", "-level2", s"f${f}_to_ui${i}"), Seq(d))
-        }.unzip
-        test(
-            s"RecF${f}ToUI${i}",
-            () => new ValExec_RecFNToUIN(exp(f), sig(f), i),
-            "RecFNToUIN.cpp",
-            softfloatArgs,
-            Some(dutArgs)
-        )
-    }
-
-    "RecF16ToUI32" should "pass" in {
-        check(test(16, 32))
-    }
-    "RecF16ToUI64" should "pass" in {
-        check(test(16, 64))
-    }
-    "RecF32ToUI32" should "pass" in {
-        check(test(32, 32))
-    }
-    "RecF32ToUI64" should "pass" in {
-        check(test(32, 64))
-    }
-    "RecF64ToUI32" should "pass" in {
-        check(test(64, 32))
-    }
-    "RecF64ToUI64" should "pass" in {
-        check(test(64, 64))
-    }
-}
-
-class RecFNToINSpec extends FMATester {
-    def test(f: Int, i: Int): Seq[String] = {
-        val (softfloatArgs, dutArgs) = roundings.map { case (s, d) =>
-            (s +: Seq("-exact", "-level2", s"f${f}_to_i${i}"), Seq(d))
-        }.unzip
-        test(
-            s"RecF${f}ToI${i}",
-            () => new ValExec_RecFNToIN(exp(f), sig(f), i),
-            "RecFNToIN.cpp",
-            softfloatArgs,
-            Some(dutArgs)
-        )
-    }
-
-    "RecF16ToI32" should "pass" in {
-        check(test(16, 32))
-    }
-    "RecF16ToI64" should "pass" in {
-        check(test(16, 64))
-    }
-    "RecF32ToI32" should "pass" in {
-        check(test(32, 32))
-    }
-    "RecF32ToI64" should "pass" in {
-        check(test(32, 64))
-    }
-    "RecF64ToI32" should "pass" in {
-        check(test(64, 32))
-    }
-    "RecF64ToI64" should "pass" in {
-        check(test(64, 64))
-    }
-}
